==> 9_lightfm.ipynb <==
import pandas as pd
import numpy as np
from lightfm import LightFM
from lightfm.data import Dataset
from lightfm.evaluation import precision_at_k
import time 

## Reading in Files

# reading in files > 3
utility_train_filtered = pd.read_parquet('utility_train_local_all_filtered.parquet')
utility_val_filtered = pd.read_parquet('utility_val_local_all_filtered.parquet')

# reading in files > 10
utility_train_filtered10 = pd.read_parquet('utility_train_local_all_filtered10.parquet')
utility_val_filtered10 = pd.read_parquet('utility_val_local_all_filtered10.parquet')

# reading in test
utility_test = pd.read_parquet('utility_test_local_all.parquet')

## Model > 10

### Train

user_id_train = utility_train_filtered10['user_id'].unique().astype('int64')
song_id_train = utility_train_filtered10['song_id'].unique().astype('int64')

data = Dataset()
data.fit(user_id_train, song_id_train)

train_values = utility_train_filtered.values.astype('int64')
train, weights_matrix_train = data.build_interactions([tuple(i) for i in train_values])

# 8 threads
start_time = time.time()
model = LightFM(loss='warp')
model.fit(train, epochs = 50, num_threads = 8)
end_time = time.time()
print("Time taken to run the model with 8 thread is", end_time - start_time)

# 1 threads
start_time = time.time()
model = LightFM(loss='warp')
model.fit(train, epochs = 50, num_threads = 1)
end_time = time.time()
print("Time taken to run the model with 1 thread is", end_time - start_time)

training_map = precision_at_k(model, train, k = 100, num_threads = 8).mean()
print("The precision at 100 of training data is", training_map)

### Test

utility_test_filtered = utility_test[utility_test['user_id'].isin(user_id_train)].reset_index(drop = True)
utility_test_filtered = utility_test_filtered[utility_test_filtered['song_id'].isin(song_id_train)]\
                        .reset_index(drop = True)

test_values = utility_test_filtered.values.astype('int64')
test, weights_matrix_test = data.build_interactions([tuple(i) for i in test_values])

# test precision
test_map = precision_at_k(model, test, k = 100, num_threads = 8).mean()
print("The precision at 100 of test data is", test_map)

## Model > 3

### Train

user_id_train = utility_train_filtered['user_id'].unique().astype('int64')
song_id_train = utility_train_filtered['song_id'].unique().astype('int64')

data = Dataset()
data.fit(user_id_train, song_id_train)

train_values = utility_train_filtered.values.astype('int64')
train, weights_matrix_train = data.build_interactions([tuple(i) for i in train_values])

# 8 threads
start_time = time.time()
model = LightFM(loss='warp')
model.fit(train, epochs = 50, num_threads = 8)
end_time = time.time()
print("Time taken to run the model with 8 thread is", end_time - start_time)

# 1 threads
start_time = time.time()
model = LightFM(loss='warp')
model.fit(train, epochs = 50, num_threads = 1)
end_time = time.time()
print("Time taken to run the model with 1 thread is", end_time - start_time)

training_map = precision_at_k(model, train, k = 100, num_threads = 8).mean()
print("The precision at 100 of training data is", training_map)

### Test

utility_test_filtered = utility_test[utility_test['user_id'].isin(user_id_train)].reset_index(drop = True)
utility_test_filtered = utility_test_filtered[utility_test_filtered['song_id'].isin(song_id_train)]\
                        .reset_index(drop = True)

test_values = utility_test_filtered.values.astype('int64')
test, weights_matrix_test = data.build_interactions([tuple(i) for i in test_values])

# test precision
test_map = precision_at_k(model, test, k = 100, num_threads = 8).mean()
print("The precision at 100 of test data is", test_map)
